requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nextcord==2.5.0
pydantic==1.10.9
pymongo==4.4.0
python-dotenv==1.0.0
typing_extensions==4.6.3
yarl==1.9.2
